@@ -1,36 +1,27 @@
 {
-  "name": "political_exam",
-  "version": "0.1.0",
+  "name": "crawlervisualization",
+  "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^1.6.2",
-    "core-js": "^3.8.3",
-    "element-ui": "^2.15.14",
-    "fuse.js": "^7.0.0",
-    "pdfmake": "^0.2.15",
-    "political_exam": "file:",
-    "vant": "^2.13.2",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2",
-    "vuex-persistedstate": "^4.1.0"
+    "@element-plus/icons-vue": "^2.3.1",
+    "element-plus": "^2.9.5",
+    "pinia": "^3.0.1",
+    "vue": "^3.5.13",
+    "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "less": "^4.0.0",
-    "less-loader": "^8.0.0",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "less": "^4.2.2",
+    "unplugin-icons": "*",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "unplugin-auto-import": "^0.19.0",
+    "unplugin-vue-components": "^0.28.0",
+    "vite": "^6.0.1",
+    "vite-plugin-vue-devtools": "^7.6.5"
+  }
 }
