@@ -1,18 +1,12 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+
 import App from './App.vue'
 import router from './router'
-import store from './store'
-import ElementUI from 'element-ui'
-import 'element-ui/lib/theme-chalk/index.css';
-import Vant from 'vant';
-import 'vant/lib/index.css';
 
-Vue.use(Vant);
-Vue.use(ElementUI)
-Vue.config.productionTip = false
+const app = createApp(App)
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+app.use(createPinia())
+app.use(router)
+
+app.mount('#app')
